+++ lib/post_frame.dart
@@ -1,6 +1,6 @@
 /// frame package
 ///
 /// This library exports all public APIs for the frame package.
-library frame;
+library;
 
 export 'src/post_frame.dart';